Validate gc_run_interval range
